+++ src/Parser/Cli.hs
@@ -1,3 +1,5 @@
+-- | Command-line parsers.
+
 module Parser.Cli
   ( module Options.Applicative
   ) where